Separate into route and controller
